requirements.txt
Flask
pandas
openpyxl
python-docx
gunicorn
